feat(react-native-kit): remove CommonDialog, ToastDialog methods from provider
